requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
dj-rest-auth==7.0.1
Django==5.2.1
django-allauth==65.9.0
django-cors-headers==4.7.0
django-seed==0.3.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Faker==37.1.0
idna==3.10
mysqlclient==2.2.7
numpy==2.2.6
opencv-python==4.11.0.86
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
requests==2.32.4
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
toposort==1.10
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
